+++ code/engine.vc2008/xrGame/ui/UILogsWnd.cpp
@@ -222,8 +222,8 @@ void CUILogsWnd::PerformWork()
 {
 	if(!m_news_in_queue.empty())
 	{
-		size_t count = std::min((u32)30, m_news_in_queue.size());
-		for(u32 i=0; i<count;++i)
+		size_t count = std::min((size_t)30, m_news_in_queue.size());
+		for(size_t i=0; i<count;++i)
 		{
 			GAME_NEWS_VECTOR& news_vector = Actor()->game_news_registry->registry().objects();
 			u32 idx						= m_news_in_queue.back();
